README.md
we planned a smart system which is trained by a neural network(R/CNN),
to assist dentists in making a medical decision. 
The AI system receives dental x-rays which are taken in the clinic. 
Because professional dentists are not always able to recognize teeth defects the system 
propose is to assist in detecting caries bone and calculus within a few seconds.